+++ lib/lightning_web/components/layouts.ex
@@ -1,0 +1,7 @@
+defmodule LightningWeb.Layouts do
+  @moduledoc false
+
+  use LightningWeb, :html
+
+  embed_templates "layouts/*"
+end