+++ lib/lightning_web/controllers/user_reset_password_html.ex
@@ -1,0 +1,7 @@
+defmodule LightningWeb.UserResetPasswordHTML do
+  @moduledoc false
+
+  use LightningWeb, :html
+
+  embed_templates "user_reset_password_html/*"
+end