+++ lib/lightning_web/endpoint.ex
@@ -23,7 +23,7 @@ defmodule LightningWeb.Endpoint do
     at: "/",
     from: :lightning,
     gzip: true,
-    only: ~w(assets fonts images favicon.ico robots.txt)
+    only: LightningWeb.static_paths()
 
   # Code reloading can be explicitly enabled under the
   # :code_reloader configuration of your endpoint.