+++ lib/lightning_web/live/project_live/index.ex
@@ -18,7 +18,7 @@ defmodule LightningWeb.ProjectLive.Index do
          socket
          |> assign(:active_menu_item, :projects)
          |> assign(current_user: socket.assigns.current_user),
-         layout: {LightningWeb.LayoutView, :settings}}
+         layout: {LightningWeb.Layouts, :settings}}
 
       {:error, :unauthorized} ->
         {:ok,