+++ lib/lightning_web/router.ex
@@ -11,7 +11,7 @@ defmodule LightningWeb.Router do
     plug :accepts, ["html"]
     plug :fetch_session
     plug :fetch_live_flash
-    plug :put_root_layout, {LightningWeb.LayoutView, :root}
+    plug :put_root_layout, {LightningWeb.Layouts, :root}
     plug :protect_from_forgery
     plug :put_secure_browser_headers
     plug :fetch_current_user