+++ lib/lightning_web/views/user_confirmation_view.ex
@@ -1,3 +1,0 @@
-defmodule LightningWeb.UserConfirmationView do
-  use LightningWeb, :view
-end