+++ lib/lightning_web/views/user_registration_view.ex
@@ -1,3 +1,0 @@
-defmodule LightningWeb.UserRegistrationView do
-  use LightningWeb, :view
-end