+++ lib/lightning_web/views/user_reset_password_view.ex
@@ -1,3 +1,0 @@
-defmodule LightningWeb.UserResetPasswordView do
-  use LightningWeb, :view
-end