+++ test/lightning_web/controllers/user_settings_controller_test.exs
@@ -84,7 +84,7 @@
 #         })
 
 #       assert redirected_to(conn) == Routes.user_settings_path(conn, :edit)
-#       assert get_flash(conn, :info) =~ "A link to confirm your email"
+#       assert Phoenix.Flash.get(conn.assigns.flash, :info) =~ "A link to confirm your email"
 #       assert Accounts.get_user_by_email(user.email)
 #     end
 
@@ -127,14 +127,14 @@
 #     } do
 #       conn = get(conn, Routes.user_settings_path(conn, :confirm_email, token))
 #       assert redirected_to(conn) == Routes.user_settings_path(conn, :edit)
-#       assert get_flash(conn, :info) =~ "Email changed successfully"
+#       assert Phoenix.Flash.get(conn.assigns.flash, :info) =~ "Email changed successfully"
 #       refute Accounts.get_user_by_email(user.email)
 #       assert Accounts.get_user_by_email(email)
 
 #       conn = get(conn, Routes.user_settings_path(conn, :confirm_email, token))
 #       assert redirected_to(conn) == Routes.user_settings_path(conn, :edit)
 
-#       assert get_flash(conn, :error) =~
+#       assert Phoenix.Flash.get(conn.assigns.flash, :error) =~
 #                "Email change link is invalid or it has expired"
 #     end
 
@@ -142,7 +142,7 @@
 #       conn = get(conn, Routes.user_settings_path(conn, :confirm_email, "oops"))
 #       assert redirected_to(conn) == Routes.user_settings_path(conn, :edit)
 
-#       assert get_flash(conn, :error) =~
+#       assert Phoenix.Flash.get(conn.assigns.flash, :error) =~
 #                "Email change link is invalid or it has expired"
 
 #       assert Accounts.get_user_by_email(user.email)