+++ CHANGELOG.md
@@ -58,7 +58,9 @@ and this project adheres to
 - Enable distributed Erlang, allowing any number of redundant Lightning nodes to
   communicate with each other.
 - Users can set up realtime alerts for a project
-- New authorization pattern accros the entire app
+- Authorization for Lightning superusers and users, and project roles
+  viewer/editor/admin/owner. See
+  [link to test file](test/lightning/permissions_test.exs) for more details.
 
 ### Changed